Refactor BaseViewCatalog into a mix-in.
